refactor: Navigation Guard를 zustand store에서 Provider로 교체
