+++ src/components/common/FooterNav.jsx
@@ -4,8 +4,8 @@ import { useState } from 'react';
 import { Community, Proposal, Reservation } from 'public/static/icons';
 
 import LoginRequiredModal from '@components/common/LoginRequiredModal';
+import { useGuardedNavigate } from '@components/common/navigation-guard/useGuardedNavigate';
 
-import { useGuardedNavigate } from '@hooks/useGuardedNavigate';
 import useTokenStore from '@stores/useTokenStore';
 
 const FooterNav = () => {
